% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRA.R
\name{calculate_total_score}
\alias{calculate_total_score}
\title{Process Expression Data and Calculate Enrichment Scores}
\usage{
calculate_total_score(expr_file, geneSet, coef_file, pos_file, neg_file)
}
\arguments{
\item{expr_file}{A string specifying the path to the validation data file (validation.txt).}

\item{geneSet}{A list of filtered gene sets. Each element in the list is a vector of gene names or indices 
representing a specific gene group.}

\item{coef_file}{A string specifying the path to the coefficients data file obtained from the `sgr_analysis_with_plots` function.}

\item{pos_file}{A string specifying the path to the positive gene set file obtained from the `sgr_analysis_with_plots` function.}

\item{neg_file}{A string specifying the path to the negative gene set file obtained from the `sgr_analysis_with_plots` function.}

\item{exprSet_file}{A string specifying the file path of the expression dataset file (e.g., TPM.txt).
The dataset should be a tab-delimited file with samples as rows and genes as columns.}

\item{model_data}{A list containing the latent variable matrix (X.latent) and the group vector (grp.vec) generated by the 
`generate_model_data` function.}

\item{X.latent}{A matrix of latent variables generated by the `generate_model_data` function.}

\item{grp.vec}{A vector of group indices generated from the `generate_model_data` function.}

\item{alpha}{The elastic net mixing parameter (alpha between 0 and 1).}

\item{sample_info_file}{A string representing the file path to the sample information file. The file should contain
at least one column named "Type" for the response variable.}

\item{fit_results}{A list containing the fit object, coefficients dataframe, positive coefficients, negative coefficients, 
and cross-validation result obtained from the `sgr_analysis_with_plots` function.}
}
\value{
A data frame containing the enrichment scores and final scores for each sample.
}
\description{
This function reads expression data, coefficients, and gene sets, applies weights to the expression data,
and calculates enrichment scores using ssGSEA.
}
\examples{
exprSet_file <- "TPM.txt"
geneSet <- readRDS("filtered_gene_sets.rds")
model_data <- generate_model_data(exprSet_file, geneSet)
X.latent <- model_data$X.latent
grp.vec <- model_data$grp.vec
alpha <- 0.95
sample_info_file <- "phenotype.txt"
fit_results <- sgr_analysis_with_plots(X.latent, sample_info_file, grp.vec, alpha, plot_dir = "my_plots")
coef_file <- fit_results$coefficients_df
pos_file <- fit_results$SGR_pos
neg_file <- fit_results$SGR_neg
expr_file <- "validation.txt"
result <- calculate_total_score(expr_file, geneSet, coef_file, pos_file, neg_file)
print(result)

}
