% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRA.R
\name{generate_model_data}
\alias{generate_model_data}
\title{Generate Model Data from Expression Data and Filtered Gene Sets}
\usage{
generate_model_data(exprSet_file, geneSet)
}
\arguments{
\item{exprSet_file}{A string specifying the file path of the expression dataset file (TPM.txt).
The dataset should be a tab-delimited file with samples as rows and genes as columns.}

\item{geneSet}{A list of filtered gene sets. Each element in the list is a vector of gene names or indices 
representing a specific gene group.}
}
\value{
A list containing:
\item{X.latent}{A matrix of latent variables created by expanding the expression data according to the gene sets.}
\item{grp.vec}{A vector indicating the group membership of each latent variable.}
}
\description{
This function processes expression data and filtered gene sets to generate latent variables and group vectors.
The expression data is processed using a set of filtered gene sets to generate a latent variable matrix and a group vector.
}
\examples{
# Load the filtered gene sets and expression dataset
geneSet <- readRDS("filtered_gene_sets.rds")
exprSet_file <- "TPM.txt"

# Generate the model data
model_data <- generate_model_data(exprSet_file, geneSet)
X.latent <- model_data$X.latent
grp.vec <- model_data$grp.vec

}
