% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRA.R
\name{sgr_analysis_with_plots}
\alias{sgr_analysis_with_plots}
\title{Perform SGR analysis with cross-validation and plots}
\usage{
sgr_analysis_with_plots(
  X.latent,
  sample_info_file,
  grp.vec,
  alpha,
  lambda_min = 0.008,
  tolerance = 1e-06,
  nfolds = 10,
  seed = 123456,
  plot_dir = "plots"
)
}
\arguments{
\item{X.latent}{A matrix of latent variables generated by the `generate_model_data` function.}

\item{sample_info_file}{A string representing the file path to the sample information file. The file should contain
at least one column named "Type" for the response variable.}

\item{grp.vec}{A vector of group indices generated from the `generate_model_data` function.}

\item{alpha}{The elastic net mixing parameter (alpha between 0 and 1).}

\item{plot_dir}{A string representing the directory where the plots will be saved. Default is "plots".}

\item{model_data}{A list containing the latent variable matrix (X.latent) and the group vector (grp.vec) generated by the 
`generate_model_data` function.}
}
\value{
A list containing the following components:
\item{fit}{The fitted SGR model object.}
\item{coefficients_df}{A data frame of coefficients for the optimal lambda.}
\item{SGR_pos}{A character vector of feature names with positive coefficients.}
\item{SGR_neg}{A character vector of feature names with negative coefficients.}
\item{pos_coefficients}{A data frame of positive coefficients for the selected features.}
\item{neg_coefficients}{A data frame of negative coefficients for the selected features.}
\item{cv_result}{The cross-validation result object containing errors and lambda selection.}
}
\description{
This function performs the SGR (Sparse Group Regularization) analysis with a given latent feature matrix,
response vector, and group vector. It includes cross-validation for tuning the model and produces
plots for cross-validation errors and coefficients. The results are saved as PDF files in the specified
plot directory.
}
\examples{
# Example usage:
model_data <- generate_model_data(exprSet_file, geneSet)
X.latent <- model_data$X.latent
grp.vec <- model_data$grp.vec
sample_info_file <- "phenotype.txt"
result <- sgr_analysis_with_plots(X.latent, sample_info_file, 
                                   grp.vec, alpha = 0.95, 
                                   plot_dir = "my_plots")

}
